split arena repo into multiple and fix routes
